NET-81: separated phases
